Add quantile function template to HistogramQuantileProvider [does not compile]
Having a ScoreFunction inside a StatisticsProvider does not fit to the
way the whole computation graph is set up. In particular, the
StatisticsProvider does not get the necessary arguments to initialize a
ScoreFunction in the constructor.

Possible fix: Create a general functor to pass to
HistogramQuantileProvider (not a ScoreFunction), instantiate it in
HistogramQuantileFunctionAffinity, and set it via a setter in
HistogramQuantileProvider.